feat(settings): 重构WebDAV同步功能并优化菜单结构
将WebDAV同步功能从独立组件迁移至数据管理页面，统一管理数据迁移与同步相关功能。
同时优化主菜单和侧边栏结构，使用动态菜单项配置提升可维护性。
默认关闭自动展开侧边栏功能，提升移动端体验。
