+++ src/components/setting.vue
@@ -10,14 +10,11 @@
           <AppSettings />
         </div>
       </el-tab-pane>
-      <el-tab-pane label="迁移" name="data">
+      <el-tab-pane label="迁移与同步" name="data">
         <DataManagement />
       </el-tab-pane>
       <el-tab-pane label="个性化" name="personalization">
       <PersonalizationSettings />
-      </el-tab-pane>
-      <el-tab-pane label="WebDAV 同步" name="webdav">
-        <WebDAVSettings />
       </el-tab-pane>
       <el-tab-pane label="关于" name="about">
         <AboutSection />
@@ -30,7 +27,6 @@
 import { ref } from 'vue';
 import AppSettings from './settingPage/AppSettings.vue';
 import DataManagement from './settingPage/DataManagement.vue';
-import WebDAVSettings from './settingPage/WebDAVSettings.vue';
 import AboutSection from './settingPage/AboutSection.vue';
 import PersonalizationSettings from './settingPage/PersonalizationSettings.vue';
